+++ src/drawables/characters/heros/states/DirectionUpState.java
@@ -23,7 +23,7 @@ public class DirectionUpState implements DirectionState{
         while (i > hero.getPosition().getY()) {
             gc.clearRect(hero.getPosition().getX(), i, widthCell, heightCell);
             i--;
-            hero.getUpSprite().drawNextSprite(gc, widthCell, heightCell, hero.getPosition().getX(), i);
+            hero.getUpSprite().drawNextSprite(gc, widthCell, heightCell, (int)hero.getPosition().getX(), i);
         }
 
     }
